Added 'On Sale' indicator for products in 'CurrentList'
Still experimental:
If products are on sale under a particular store, the UI will indicate that to user.
In *my* db, I added an 'on_sale' (bool) field, and 'prev_price' (num) field under 'stores_carrying'. Need to discuss this in group meeting to see if that'll be the way to go. But otherwise, I can change my code to adapt.